Refactor t_trxstatus
Replace old struct with class. Replace free functions with member
functions. Increase memory safety. Make access to available file objects
explicit through use of optional instead of old instantiations.

Remove all use of deprecated read_first_x/read_next_x functions.

+++ api/legacy/include/gromacs/trajectoryanalysis/analysissettings.h
@@ -161,7 +161,7 @@ public:
      */
     bool hasRmPBC() const;
     //! Returns the currently set frame flags.
-    int frflags() const;
+    size_t frflags() const;
 
     /*! \brief
      * Sets flags.
@@ -228,7 +228,7 @@ public:
      * forces), it can call this function to load additional information
      * from the trajectory.
      */
-    void setFrameFlags(int frflags);
+    void setFrameFlags(size_t frflags);
 
     //! \copydoc ICommandLineOptionsModuleSettings::setHelpText()
     void setHelpText(const ArrayRef<const char* const>& help);
